code cleaning
git-svn-id: http://svn.miranda-ng.org/main/trunk@12457 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/FavContacts/src/contact_cache.cpp
@@ -169,8 +169,8 @@ TCHAR *nb_stristr(TCHAR *str, TCHAR *substr)
 	TCHAR *str_up = NEWTSTR_ALLOCA(str);
 	TCHAR *substr_up = NEWTSTR_ALLOCA(substr);
 
-	CharUpperBuff(str_up, mir_tstrlen(str_up));
-	CharUpperBuff(substr_up, mir_tstrlen(substr_up));
+	CharUpperBuff(str_up, (DWORD)mir_tstrlen(str_up));
+	CharUpperBuff(substr_up, (DWORD)mir_tstrlen(substr_up));
 
 	TCHAR *p = _tcsstr(str_up, substr_up);
 	return p ? (str + (p - str_up)) : NULL;
